Refactor map module into modular structure
