+++ lib/map/types.ts
@@ -1,0 +1,23 @@
+import type { TileSubtype, RoomId } from "./constants";
+import type { PlainEnemy } from "../enemy";
+import type { EnvironmentId } from "../environment";
+
+export interface MapData {
+  tiles: number[][];
+  subtypes: number[][][];
+  environment?: EnvironmentId;
+}
+
+export interface RoomSnapshot {
+  mapData: MapData;
+  entryPoint: [number, number];
+  enemies?: PlainEnemy[];
+  potOverrides?: Record<string, TileSubtype.FOOD | TileSubtype.MED>;
+}
+
+export interface RoomTransition {
+  from: RoomId;
+  to: RoomId;
+  position: [number, number];
+  targetEntryPoint?: [number, number];
+}
